+++ screens/AddPortScreen.tsx
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+export default function AddPortScreen() {
+  return (
+    <View>
+      <Text>AddPortScreen</Text>
+    </View>
+  )
+}
