+++ screens/NotificationScreen.tsx
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+export default function NotificationScreen() {
+  return (
+    <View>
+      <Text>NotificationScreen</Text>
+    </View>
+  )
+}
