+++ screens/ProfileScreen.tsx
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+export default function ProfileScreen() {
+  return (
+    <View>
+      <Text>ProfileScreen</Text>
+    </View>
+  )
+}
